AdminGelControllerTest( getForm and postForm method is tested)

+++ src/main/java/org/george/jylishop/controllers/AdminGelController.java
@@ -69,7 +69,7 @@ public class AdminGelController {
                                  @RequestParam String title,
                                  @RequestParam Double volume,
                                  @RequestParam Double price,
-                                 @RequestParam Double reactantPercent,
+                                 @RequestParam Integer reactantPercent,
                                  @RequestParam String description,
                                  @RequestParam String picture) {
